Fixes: Create action button component & refactoring
