+++ src/views/dashboard/UsageDashboardChart.vue
@@ -57,7 +57,3 @@ export default {
   }
 }
 </script>
-
-<style scoped>
-
-</style>
